+++ packages/generator/src/edmx-to-vdm/edmx-to-vdm-util.ts
@@ -1,14 +1,18 @@
 import { createLogger, last } from '@sap-cloud-sdk/util';
-import { EdmxMetadata } from '../edmx-parser/edmx-file-reader';
-import { EdmxProperty } from '../edmx-parser/common/edmx-types';
+import type { EdmxProperty } from '../edmx-parser/common/edmx-types';
+import type { EdmxMetadata } from '../edmx-parser/edmx-file-reader';
+import type { EdmxFunctionImportV2 } from '../edmx-parser/v2/edm-types';
 import {
   edmToFieldType,
   edmToTsType,
   getFallbackEdmTypeIfNeeded
 } from '../generator-utils';
-import { VdmComplexType, VdmEnumType, VdmMappedEdmType } from '../vdm-types';
-import { EdmxOperation } from '../edmx-parser/v4/edm-types';
-import { EdmxFunctionImportV2 } from '../edmx-parser/v2/edm-types';
+import type {
+  VdmComplexType,
+  VdmEnumType,
+  VdmMappedEdmType
+} from '../vdm-types';
+import type { EdmxJoinedOperation } from './v4';
 
 const logger = createLogger({
   package: 'generator',
@@ -193,7 +197,7 @@ export const propertyJsType = (type: string): string | undefined =>
  * @internal
  */
 export function hasUnsupportedParameterTypes(
-  operation: EdmxOperation | EdmxFunctionImportV2
+  operation: EdmxJoinedOperation | EdmxFunctionImportV2
 ): boolean {
   const unsupportedParameters = operation.Parameter.filter(
     p => !isEdmType(p.Type)