+++ packages/generator/src/edmx-to-vdm/v2/entity.ts
@@ -65,7 +65,9 @@ export function generateEntitiesV2(
       associations,
       classNames,
       formatter
-    )
+    ),
+    functions: [],
+    actions: []
   }));
 }