+++ packages/generator/src/entity/imports.spec.ts
@@ -3,7 +3,7 @@ import { entityImportDeclarations, otherEntityImports } from './imports';
 
 describe('imports', () => {
   it('importDeclarations', () => {
-    const actual = entityImportDeclarations(breakfastEntity, 'v2');
+    const actual = entityImportDeclarations(breakfastEntity, foodService, 'v2');
 
     expect(
       actual.map(imports => ({
@@ -12,6 +12,14 @@ describe('imports', () => {
         isTypeOnly: !!imports.isTypeOnly
       }))
     ).toEqual(expectedEntityImports);
+  });
+
+  it('importDeclarations oData v4', () => {
+    const actual = entityImportDeclarations(breakfastEntity, foodService, 'v4');
+
+    expect(actual.map(imports => imports.namedImports)).toEqual(
+      expectedEntityImportsV4[0].namedImports
+    );
   });
 
   it('otherEntityImports', () => {
@@ -33,6 +41,34 @@ const expectedEntityImports = [
       'DefaultDeSerializers',
       'DeSerializers',
       'DeserializedType'
+    ],
+    isTypeOnly: false
+  },
+  {
+    namedImports: ['BreakfastApi'],
+    moduleSpecifier: './BreakfastApi',
+    isTypeOnly: true
+  }
+];
+
+const expectedEntityImportsV4 = [
+  {
+    moduleSpecifier: '@sap-cloud-sdk/odata-v4',
+    namedImports: [
+      ['Entity', 'DefaultDeSerializers', 'DeSerializers', 'DeserializedType'],
+      ['BreakfastApi'],
+      [
+        'edmToTs',
+        'transformReturnValueForEdmType',
+        'DeSerializers',
+        'DefaultDeSerializers',
+        'defaultDeSerializers',
+        'ActionImportParameter',
+        'ActionImportRequestBuilder',
+        'FunctionImportParameter',
+        'FunctionImportRequestBuilder'
+      ],
+      ['foodService']
     ],
     isTypeOnly: false
   },