+++ packages/odata-common/src/request-builder/batch/batch-request-serializer.ts
@@ -1,16 +1,17 @@
 import { unixEOL } from '@sap-cloud-sdk/util';
 import voca from 'voca';
-import { v4 as uuid } from 'uuid';
 import { ODataRequest } from '../../request/odata-request';
 import { ODataRequestConfig } from '../../request/odata-request-config';
 import { MethodRequestBuilder } from '../request-builder-base';
 import { DeSerializers } from '../../de-serializers';
-import { BatchChangeSet } from './batch-change-set';
+import { EntityBase } from '../../entity-base';
+import { WithBatchReference } from '../../request';
+import type { BatchRequestBuilder } from './batch-request-builder';
 import {
   BatchRequestSerializationOptions,
   BatchSubRequestPathType
 } from './batch-request-options';
-import type { BatchRequestBuilder } from './batch-request-builder';
+import { BatchChangeSet } from './batch-change-set';
 
 /**
  * Serialize change set to string.
@@ -43,8 +44,13 @@ export function serializeChangeSet<DeSerializersT extends DeSerializers>(
  * @returns The serialized string representation of a multipart request, including the multipart headers.
  * @internal
  */
-export function serializeRequest(
-  request: Omit<MethodRequestBuilder, 'execute'>,
+export function serializeRequest<
+  EntityT extends EntityBase,
+  DeSerializersT extends DeSerializers,
+  ReturnT,
+  ParametersT
+>(
+  request: AllRequestBuilders<EntityT, DeSerializersT>,
   options: BatchRequestSerializationOptions = {}
 ): string {
   const odataRequest = new ODataRequest(
@@ -65,7 +71,9 @@ export function serializeRequest(
   return [
     'Content-Type: application/http',
     'Content-Transfer-Encoding: binary',
-    ...(method !== 'GET' ? [`Content-Id: ${uuid()}`] : []),
+    ...(method !== 'GET' && isRequestBuilderWithBatchReference(request)
+      ? [`Content-Id: ${request.getBatchReference().id}`]
+      : []),
     '',
     `${method} ${getUrl(odataRequest, options.subRequestPathType)} HTTP/1.1`,
     ...(requestHeaders.length ? requestHeaders : ['']),
@@ -73,6 +81,23 @@ export function serializeRequest(
     ...getPayload(request),
     ''
   ].join(unixEOL);
+}
+
+type RequestBuildersWithBatchReference = Omit<MethodRequestBuilder, 'execute'> &
+  WithBatchReference;
+
+type AllRequestBuilders<
+  EntityT extends EntityBase,
+  DeSerializersT extends DeSerializers
+> = RequestBuildersWithBatchReference | Omit<MethodRequestBuilder, 'execute'>;
+
+function isRequestBuilderWithBatchReference<
+  EntityT extends EntityBase,
+  DeSerializersT extends DeSerializers
+>(
+  request: AllRequestBuilders<EntityT, DeSerializersT>
+): request is RequestBuildersWithBatchReference {
+  return !!(request as WithBatchReference).getBatchReference;
 }
 
 function getUrl<ConfigT extends ODataRequestConfig>(