+++ packages/odata-common/src/request-builder/update-request-builder-base.ts
@@ -3,6 +3,7 @@ import {
   transformVariadicArgumentToArray
 } from '@sap-cloud-sdk/util';
 import { HttpResponse } from '@sap-cloud-sdk/http-client';
+import { v4 as uuid } from 'uuid';
 import { EntityBase, EntityIdentifiable } from '../entity-base';
 import { extractEtagFromHeader } from '../entity-deserializer';
 import { EntitySerializer } from '../entity-serializer';
@@ -12,6 +13,10 @@ import { ODataUri } from '../uri-conversion';
 import { Selectable } from '../selectable';
 import { DeSerializers } from '../de-serializers/de-serializers';
 import { EntityApi } from '../entity-api';
+import {
+  BatchReference,
+  WithBatchReference
+} from '../request/odata-request-traits';
 import { MethodRequestBuilder } from './request-builder-base';
 
 /**
@@ -25,11 +30,12 @@ export abstract class UpdateRequestBuilderBase<
   extends MethodRequestBuilder<
     ODataUpdateRequestConfig<EntityT, DeSerializersT>
   >
-  implements EntityIdentifiable<EntityT, DeSerializersT>
+  implements EntityIdentifiable<EntityT, DeSerializersT>, WithBatchReference
 {
   readonly _deSerializers: DeSerializersT;
   private ignored: Set<string>;
   private required: Set<string>;
+  private _batchReference: BatchReference;
 
   /**
    * Creates an instance of UpdateRequestBuilder.
@@ -67,6 +73,25 @@ export abstract class UpdateRequestBuilderBase<
 
   get entity(): EntityT {
     return this._entity;
+  }
+
+  /**
+   * Gets identifier for the batch request.
+   * @returns Batch request identifier.
+   */
+  getBatchReference(): BatchReference {
+    if (!this._batchReference) {
+      this.setBatchId(uuid());
+    }
+    return this._batchReference;
+  }
+
+  /**
+   * Sets user-defined identifier for the batch reference.
+   * @param id - User-defined batch reuest identifier.
+   */
+  setBatchId(id: string): void {
+    this._batchReference = { id };
   }
 
   /**