+++ packages/odata-v4/src/request-builder/index.ts
@@ -9,3 +9,5 @@ export * from './response-data-accessor';
 export * from './response-transformers';
 export * from './update-request-builder';
 export * from './batch-request-builder';
+export * from './bound-action-import-request-builder';
+export * from './bound-function-import-request-builder';