+++ test-packages/e2e-tests/srv/test-service.js
@@ -6,7 +6,7 @@ module.exports = async srv => {
   cds.model = cds.compile.for.nodejs(csn);
 
   const db = await cds.connect.to('db');
-  const { TestEntity } = srv.entities;
+  const { TestEntity, TestEntityWithMultipleKeys } = srv.entities;
 
   // bound function
   srv.on('getStringProperty', 'TestEntity', async oRequest => {
@@ -15,6 +15,49 @@ module.exports = async srv => {
       .run(SELECT.one.from(oRequest.query.SELECT.from));
     oRequest.reply(entity.StringProperty);
   });
+
+  srv.on(
+    'boundFunctionWithoutArgumentsWithMultipleKeys',
+    'TestEntityWithMultipleKeys',
+    async oRequest => {
+      oRequest.reply('boundFunctionWithoutArgumentsWithMultipleKeys Result Value');
+    }
+  );
+
+  srv.on(
+    'boundActionWithoutArgumentsWithMultipleKeys',
+    'TestEntityWithMultipleKeys',
+    async oRequest => {
+      oRequest.reply('boundActionWithoutArgumentsWithMultipleKeys Result Value');
+    }
+  );
+
+  srv.on('boundFunctionWithoutArguments', 'TestEntity', async oRequest => {
+    oRequest.reply('boundFunctionWithoutArguments Result Value');
+  });
+
+  srv.on('boundFunctionWithArguments', 'TestEntity', async oRequest => {
+    oRequest.reply(`boundFunctionWithArguments ${oRequest.data.param1} ${oRequest.data.param2} Result Value`);
+  });
+
+  srv.on(
+    'boundFunctionWithArgumentsWithMultipleKeys',
+    'TestEntityWithMultipleKeys',
+    async oRequest => {
+      oRequest.reply(`boundFunctionWithArgumentsWithMultipleKeys ${oRequest.data.param1} ${oRequest.data.param2} Result Value`);
+    }
+  );
+
+  srv.on(
+    'boundFunctionWithoutArgumentsComplexReturnType',
+    'TestEntity',
+    async oRequest => {
+      oRequest.reply({
+        someMessage: 'boundFunctionWithoutArgumentsComplexReturnType Result Value',
+        someId: 42
+      });
+    }
+  );
 
   // bound action
   srv.on('deleteEntity', 'TestEntity', async oRequest => {
@@ -25,6 +68,10 @@ module.exports = async srv => {
       .transaction(oRequest)
       .run(DELETE.from(TestEntity).byKey(entity.KeyTestEntity));
     oRequest.reply(entity.KeyTestEntity);
+  });
+
+  srv.on('boundActionWithoutArguments', 'TestEntity', async oRequest => {
+    oRequest.reply('boundActionWithoutArguments Result Value');
   });
 
   // unbound function
@@ -51,6 +98,18 @@ module.exports = async srv => {
     const entity = await SELECT.one
       .from(TestEntity)
       .where({ KeyTestEntity: param });
+    oRequest.reply(entity);
+  });
+
+  srv.on('getByKeyWithMultipleKeys', async oRequest => {
+    const param1 = oRequest.data.param1;
+    const param2 = oRequest.data.param2;
+    const param3 = oRequest.data.param3;
+    const entity = await SELECT.one.from(TestEntityWithMultipleKeys).where({
+      KeyTestEntityWithMultipleKeys: param1,
+      StringPropertyWithMultipleKeys: param2,
+      BooleanPropertyWithMultipleKeys: param3
+    });
     oRequest.reply(entity);
   });