+++ packages/connectivity/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-cloud-sdk/connectivity
 
+## 2.11.0
+
+### Patch Changes
+
+- @sap-cloud-sdk/util@2.11.0
+
 ## 2.10.0
 
 ### Patch Changes